wip metrics modal

+++ pkg/graph/graph.go
@@ -1,17 +1,17 @@
 package graph
 
-func New(n int) Graph {
-	cap := make([]int, n*n)
-	cost := make([]int, n*n)
+func New(n int64) Graph {
+	cap := make([]int64, n*n)
+	cost := make([]int64, n*n)
 	return Graph{n, cap, cost}
 }
 
 type Graph struct {
-	n         int
-	cap, cost []int
+	n         int64
+	cap, cost []int64
 }
 
-func (g *Graph) AddEdge(s, t, cap, cost int) {
+func (g *Graph) AddEdge(s, t, cap, cost int64) {
 	g.cap[s*g.n+t] = cap
 	g.cap[t*g.n+s] = cap
 	g.cost[s*g.n+t] = cost
