+++ pkg/graph/min_cost_max_flow_test.go
@@ -48,16 +48,16 @@ func BenchmarkMinCostMaxFlow(b *testing.B) {
 func TestMinCostMaxFlowLarge(t *testing.T) {
 	g := New(201)
 
-	for i := 1; i < 100; i++ {
+	for i := int64(1); i < 100; i++ {
 		g.AddEdge(0, i, 1, 0)
 	}
-	for i := 100; i < 200; i++ {
+	for i := int64(100); i < 200; i++ {
 		g.AddEdge(i, 200, 1, 0)
 	}
 
 	r := rand.New(rand.NewSource(0))
-	for i := 1; i < 100; i++ {
-		for j := 100; j < 200; j++ {
+	for i := int64(1); i < 100; i++ {
+		for j := int64(100); j < 200; j++ {
 			if r.Float32() < 0.1 {
 				g.AddEdge(i, j, 1, 1)
 			}
